fix(apple/macOS): Fetch providerStopReason over IPC on macOS (#7570)
Since we no longer have access to the tunnel's group container from the
app process, we need to move reading the last VPN stop reason to an IPC
call. This call is used to show an alert to the user when their connlib
session has been disconnected due to 401: Unauthorized.

Fixes #7468