fix(gui-client): don't fail on repeated deep-links (#7572)
Firezone's authentication scheme uses deep-links to transfer the secret
token via the login-flow using the browser to the application. Such a
deep-link can be opened multiple times, even if we are already signed
in. In such a case, and in any other where we don't have a pending
sign-in request, we currently generate an error.

This is unnecessary as we can simply discard the token received from the
deep-link.