fix(apple): Load tunnel manager after creating it (#7593)
When launching Firezone for the first time, the VPN profile doesn't
exist. We prompt the user to create it with "Grant VPN Permission", but
then we fail to reload it, which initializes the tunnelManager instance
variables properly and binds observers.

The result of this was that we failed to react to VPN status changes on
the first launch of Firezone.

This can (and should be) refactored to be cleaner, but that is out of
scope for this PR and will be saved for #6554.


Refs #7579