refactor(apple): Clean up use of DispatchQueue.main (#7612)
In several places we were unnecessarily wrapping code in either a `Task`
or `DispatchQueue.main.async` block.

The former is only needed if the API we're calling itself exposes an
`async` function or we're running some long-running task.

The latter is typically used for UI updates.

At all other times it is preferable to simply use blocking functions and
let the caller of these best determine how to execute them.