fix(apple): Move status observer setup to `init()` (#7673)
This is a very minor regression caused by #7555. We didn't bind the
status update observer until after the VPN profile was created. In
practice this didn't cause an issue because the status never changes
until a VPN profile is created, but still thought it's good to fix.

The JSON encoder is now configured in the `init()` as well like any
other instance variable should be.